Move Fabric-specific StaffItem functionality to Fabric project
